+++ app/src/main/java/app/eluvio/wallet/screens/dashboard/mymedia/MyMedia.kt
@@ -1,6 +1,7 @@
 package app.eluvio.wallet.screens.dashboard.mymedia
 
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
@@ -9,9 +10,11 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.key.Key
 import androidx.compose.ui.input.key.onPreviewKeyEvent
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -23,10 +26,12 @@ import androidx.tv.foundation.lazy.list.items
 import androidx.tv.foundation.lazy.list.rememberTvLazyListState
 import androidx.tv.material3.MaterialTheme
 import androidx.tv.material3.Text
+import app.eluvio.wallet.R
 import app.eluvio.wallet.data.entities.MediaEntity
 import app.eluvio.wallet.navigation.DashboardTabsGraph
 import app.eluvio.wallet.navigation.LocalNavigator
 import app.eluvio.wallet.navigation.asPush
+import app.eluvio.wallet.screens.common.EluvioLoadingSpinner
 import app.eluvio.wallet.screens.common.MediaItemCard
 import app.eluvio.wallet.screens.common.spacer
 import app.eluvio.wallet.screens.dashboard.myitems.AllMediaProvider
@@ -50,6 +55,22 @@ fun MyMedia() {
 
 @Composable
 private fun MyMedia(state: MyMediaViewModel.State) {
+    Box(
+        contentAlignment = Alignment.Center,
+        modifier = Modifier.fillMaxSize()
+    ) {
+        if (state.loading) {
+            EluvioLoadingSpinner()
+        } else if (state.isEmpty()) {
+            Text(stringResource(R.string.no_content_warning))
+        } else {
+            MyMediaGrid(state)
+        }
+    }
+}
+
+@Composable
+private fun MyMediaGrid(state: MyMediaViewModel.State) {
     val scrollState = rememberTvLazyListState()
     val scope = rememberCoroutineScope()
     TvLazyColumn(
@@ -166,5 +187,5 @@ private val listSpacerSize = 28.dp
 @Composable
 @Preview(device = Devices.TV_720p)
 private fun MyMediaPreview() = EluvioThemePreview {
-    MyMedia(MyMediaViewModel.State())
+    MyMedia(MyMediaViewModel.State(loading = false))
 }